Experiment/DeepSeek-RC/111/requirements.txt
pandas      
openpyxl==3.1.2          
pydantic==2.5.2          

requests==2.31.0         
openai==1.3.7            

torch==2.6.0             
transformers==4.35.2     
datasets==2.14.5         
peft==0.6.0              
accelerate==0.24.1       
sentencepiece==0.1.99    
protobuf==3.20.3         
tqdm==4.66.1         
numpy
python-dotenv